# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PDF Editor Backend API",
    "description": "API for authentication, user management, and history tracking",
    "version": "1.0.0",
    "contact": {
      "email": "admin@pdfeditor.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "PDF Editor Backend Service"
    }
  ],
...
  "paths": {
...
    "/history/log": {
      "post": {
        "tags": ["history"],
        "summary": "Log activity",
        "description": "Record a new activity in the history log (for frontend operations)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "description": "The action performed",
                    "example": "pdf-merge"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the action",
                    "example": "Merged 3 PDF files"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Additional metadata about the operation"
                  }
                },
                "required": ["action"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Activity logged successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Activity logged successfully"
                    },
                    "historyId": {
                      "type": "integer",
                      "example": 123
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Action is required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Failed to log activity"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/pdfLogs/log": {
      "post": {
        "tags": ["pdf-logs"],
        "summary": "Log PDF operation",
        "description": "Record a PDF operation in the logs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "description": "The action performed (without pdf- prefix)",
                    "example": "merge"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the operation",
                    "example": "Merged 3 PDF files"
                  },
                  "fileId": {
                    "type": "string",
                    "description": "ID of the PDF file",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  },
                  "fileName": {
                    "type": "string",
                    "description": "Name of the PDF file",
                    "example": "document.pdf"
                  },
                  "operationType": {
                    "type": "string",
                    "description": "Type of operation",
                    "example": "merge"
                  }
                },
                "required": ["action"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation logged successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Operation logged successfully"
                    },
                    "historyId": {
                      "type": "integer",
                      "example": 123
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Action is required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "API key required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Failed to log operation"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/pdfLogs": {
      "get": {
        "tags": ["pdf-logs"],
        "summary": "Get PDF operation history",
        "description": "Get PDF operation history for the authenticated user",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of records per page"
          }
        ],
        "responses": {
          "200": {
            "description": "PDF operation history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 123
                          },
                          "userId": {
                            "type": "integer",
                            "example": 456
                          },
                          "action": {
                            "type": "string",
                            "example": "pdf-merge"
                          },
                          "description": {
                            "type": "string",
                            "example": "Merged 3 PDF files"
                          },
                          "ipAddress": {
                            "type": "string",
                            "example": "192.168.1.1"
                          },
                          "city": {
                            "type": "string",
                            "example": "Bratislava"
                          },
                          "country": {
                            "type": "string",
                            "example": "Slovakia"
                          },
                          "accessType": {
                            "type": "string",
                            "enum": ["frontend", "api"],
                            "example": "api"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-05-10T15:30:00.000Z"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "example": 50
                        },
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "pages": {
                          "type": "integer",
                          "example": 5
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Authentication required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Failed to fetch history"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/pdfLogs/check": {
      "get": {
        "tags": ["pdf-logs"],
        "summary": "Check PDF logs",
        "description": "Check if PDF logs exist (for debugging)",
        "responses": {
          "200": {
            "description": "PDF logs check completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "example": 150
                    },
                    "pdfOperations": {
                      "type": "integer",
                      "example": 75
                    },
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string",
                      "example": "History check completed successfully"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Failed to check history records"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}